requirements.txt
keyboard>=0.13.5
pystray>=0.19.5
Pillow>=10.0.0
pyperclip>=1.8.2
pyautogui>=0.9.54
requests>=2.31.0
